functions/users/requirements.txt
flask==3.1.1
functions-framework==3.8.3
google-cloud-firestore>=2.0.0